Add ID to BasicInfo (#2)
Adds the 'id' string field to the BasicInfo message. Note that
this change is backwards-incompatible because it changes the
version associated with the 'version' field. This is acceptable
because the API is still in early development, but can cause
confusing errors when running with inconsistent .proto definitions
between server and client.